[HM-API2][QAPI-1230] Migrate homematic binding to ESH
Apply qivicon specific modifications to the binding. These are eiter
dependencies which are different on qivicon, or adaptions that are
valuable for qivicon but not for ESH.
The changes were taken from homematic-binding repository. Original
authors were Julian Dax, Florian Stolte, Michael Reitler

QAPI-983: [HM-API2] Provide correct default label
HM-API2: Replace label of DiscoveryResult with empty String

The labels for discovered devices in the Pairing Wizard 2 should be
taken from PIS for the homematic binding. Therefore, the label of every
DiscoveryResult must be set to the empty string.

[HM-API2] QAPI-1153 PoC Security

Allows the security key to be set on a homematic bridge via the
HomematicBridgeHandler

Initial integration
* adjust build.properties and MANIFEST.MF
* Add gitignore
* Add libraries